Move setTouchInteraction to a protocol extension.
