+++ Sources/SwiftUICharts/PieChart/Models/ChartData/PieChartData.swift
@@ -68,14 +68,6 @@ public final class PieChartData: CTPieChartDataProtocol {
         self.makeDataPoints()
     }
     
-    // MARK: Touch
-    public func setTouchInteraction(touchLocation: CGPoint, chartSize: CGRect) {
-        self.infoView.isTouchCurrent   = true
-        self.infoView.touchLocation    = touchLocation
-        self.infoView.chartSize        = chartSize
-        self.getDataPoint(touchLocation: touchLocation, chartSize: chartSize)
-    }
-    
     public func getTouchInteraction(touchLocation: CGPoint, chartSize: CGRect) -> some View { EmptyView() }
 
     public typealias Set       = PieDataSet
@@ -85,7 +77,7 @@ public final class PieChartData: CTPieChartDataProtocol {
 
 // MARK: - Touch
 extension PieChartData: TouchProtocol {
-    internal func getDataPoint(touchLocation: CGPoint, chartSize: CGRect) {
+    public func getDataPoint(touchLocation: CGPoint, chartSize: CGRect) {
         var points : [PieChartDataPoint] = []
         let touchDegree = degree(from: touchLocation, in: chartSize)
                 
@@ -94,6 +86,9 @@ extension PieChartData: TouchProtocol {
             points.append(data)
         }
         self.infoView.touchOverlayInfo = points
+    }
+    public func getPointLocation(dataSet: PieDataSet, touchLocation: CGPoint, chartSize: CGRect) -> CGPoint? {
+        return nil
     }
 }
