+++ Sources/SwiftUICharts/PieChart/Models/Protocols/PieChartProtocolsExtentions.swift
@@ -69,12 +69,6 @@ extension CTPieDoughnutChartDataProtocol where Set == MultiPieDataSet, DataPoint
     }
 }
 
-extension CTPieDoughnutChartDataProtocol {
-    internal func getPointLocation(dataSet: PieDataSet, touchLocation: CGPoint, chartSize: CGRect) -> CGPoint? {
-        return nil
-    }
-}
-
 extension CTPieDoughnutChartDataProtocol where Set == PieDataSet, DataPoint == PieChartDataPoint {
 
     /**
