+++ Sources/SwiftUICharts/SharedLineAndBar/Extras/LineAndBarEnums.swift
@@ -69,6 +69,22 @@ public enum DisplayValue {
     case center(specifier: String)
 }
 
+/**
+ Where to start drawing the line chart from.
+ ```
+ case minimumValue // Lowest value in the data set(s)
+ case minimumWithMaximum(of: Double) // Set a custom baseline
+ case zero // Set 0 as the lowest value
+ ```
+ */
+public enum Baseline {
+    /// Lowest value in the data set(s)
+    case minimumValue
+    /// Set a custom baseline
+    case minimumWithMaximum(of: Double)
+    /// Set 0 as the lowest value
+    case zero
+}
 
 /**
  Where to end drawing the chart.
